refactor: Achieve 10/10 REST API compliance - schema cleanup
This commit completes the REST API compliance work by addressing the two
minor issues identified in the initial audit, achieving a perfect 10/10 score.

Changes:
- Moved password reset schemas from router to dedicated schema file
  - CreatePasswordResetRequest, VerifyResetTokenResponse, CompletePasswordResetRequest
  - Now properly organized in src/schemas/auth.py
- Removed duplicate OAuth router (src/api/v1/auth.py)
  - Kept modern implementation: provider_authorization.py
  - Removed legacy implementation to eliminate confusion
- Removed deprecated test file (tests/api/test_auth_endpoints.py)
  - Tests for legacy OAuth router are no longer needed
  - Modern tests remain in test_provider_authorization_endpoints.py
- Updated router registry (src/api/v1/__init__.py)
  - Removed registration of deprecated auth.py router

Impact:
- ✅ Zero inline schemas in API routers
- ✅ No duplicate or conflicting router implementations
- ✅ Complete separation of concerns
- ✅ All 295 tests passing (76% coverage)
- ✅ Code passes lint and format checks

Documentation:
- Added comprehensive audit reports documenting 10/10 compliance
- docs/development/reviews/REST_API_AUDIT_REPORT.md (initial)
- docs/development/reviews/REST_API_AUDIT_REPORT_2025-10-05.md (final)

REST Compliance Score: 9.5/10 → 10/10 ✅

Refs: Previous REST API compliance work (Phases 1-3)
