+++ src/api/v1/password_resets.py
@@ -10,82 +10,22 @@ This design is more REST-compliant than action-oriented URLs like
 """
 
 import logging
-from typing import Optional
 
 from fastapi import APIRouter, Depends, HTTPException, status
-from pydantic import BaseModel, EmailStr, Field
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.core.database import get_session
+from src.schemas.auth import (
+    CompletePasswordResetRequest,
+    CreatePasswordResetRequest,
+    VerifyResetTokenResponse,
+)
 from src.schemas.common import MessageResponse
 from src.services.auth_service import AuthService
 
 logger = logging.getLogger(__name__)
 
 router = APIRouter()
-
-
-# Request/Response Schemas
-
-
-class CreatePasswordResetRequest(BaseModel):
-    """Request to create a password reset.
-
-    Attributes:
-        email: Email address to send reset link to.
-    """
-
-    email: EmailStr = Field(
-        ..., description="Email address to send password reset link"
-    )
-
-    model_config = {"json_schema_extra": {"example": {"email": "john.doe@example.com"}}}
-
-
-class VerifyResetTokenResponse(BaseModel):
-    """Response for token verification.
-
-    Attributes:
-        valid: Whether the token is valid.
-        email: Email address associated with token (only if valid).
-        expires_at: Token expiration timestamp (only if valid).
-    """
-
-    valid: bool = Field(..., description="Whether the token is valid")
-    email: Optional[str] = Field(
-        default=None, description="Email address associated with token"
-    )
-    expires_at: Optional[str] = Field(
-        default=None, description="Token expiration timestamp (ISO 8601)"
-    )
-
-    model_config = {
-        "json_schema_extra": {
-            "example": {
-                "valid": True,
-                "email": "john.doe@example.com",
-                "expires_at": "2025-10-05T12:00:00Z",
-            }
-        }
-    }
-
-
-class CompletePasswordResetRequest(BaseModel):
-    """Request to complete password reset.
-
-    Attributes:
-        new_password: New password meeting strength requirements.
-    """
-
-    new_password: str = Field(
-        ...,
-        description="New password (min 8 chars, 1 uppercase, 1 lowercase, 1 digit, 1 special)",
-        min_length=8,
-        max_length=128,
-    )
-
-    model_config = {"json_schema_extra": {"example": {"new_password": "NewSecure123!"}}}
-
 
 # Endpoints
