--- tests/api/test_auth_endpoints.py
@@ -1,387 +1,0 @@
-"""Tests for OAuth authentication flow endpoints.
-
-This module tests the complete OAuth flow including:
-- Authorization URL generation
-- OAuth callback handling
-- Token storage and management
-- Token refresh
-- Provider disconnection
-"""
-
-from unittest.mock import AsyncMock, patch, MagicMock
-from uuid import uuid4
-from datetime import datetime, timezone
-
-from fastapi import status
-from fastapi.testclient import TestClient
-
-from src.models.provider import Provider, ProviderConnection, ProviderToken
-
-
-class TestGetAuthorizationURL:
-    """Test suite for get_authorization_url endpoint."""
-
-    def test_get_authorization_url_success(
-        self, client: TestClient, test_provider: Provider
-    ):
-        """Test successfully getting authorization URL."""
-        with patch(
-            "src.api.v1.auth.ProviderRegistry.create_provider_instance"
-        ) as mock_registry:
-            # Mock provider instance
-            mock_provider_impl = MagicMock()
-            mock_provider_impl.get_auth_url.return_value = (
-                "https://provider.com/oauth/authorize?client_id=test"
-            )
-            mock_registry.return_value = mock_provider_impl
-
-            response = client.get(f"/api/v1/auth/{test_provider.id}/authorize")
-
-            assert response.status_code == status.HTTP_200_OK
-            data = response.json()
-            assert "auth_url" in data
-            assert "https://provider.com/oauth/authorize" in data["auth_url"]
-            assert "message" in data
-            mock_provider_impl.get_auth_url.assert_called_once()
-
-    def test_get_authorization_url_provider_not_found(self, client: TestClient):
-        """Test getting authorization URL for non-existent provider."""
-        fake_uuid = uuid4()
-        response = client.get(f"/api/v1/auth/{fake_uuid}/authorize")
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
-        assert "Provider not found" in response.json()["detail"]
-
-    def test_get_authorization_url_invalid_provider_key(
-        self, client: TestClient, test_provider: Provider
-    ):
-        """Test handling of invalid provider key."""
-        with patch(
-            "src.api.v1.auth.ProviderRegistry.create_provider_instance"
-        ) as mock_registry:
-            mock_registry.side_effect = ValueError("Invalid provider key")
-
-            response = client.get(f"/api/v1/auth/{test_provider.id}/authorize")
-
-            assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
-class TestOAuthCallback:
-    """Test suite for handle_oauth_callback endpoint."""
-
-    def test_callback_success(
-        self, client: TestClient, test_provider: Provider, db_session
-    ):
-        """Test successful OAuth callback with token exchange.
-
-        Note: Due to async/sync test infrastructure, this tests the flow
-        but may encounter session issues in the test environment.
-        """
-        # Create connection for provider
-        connection = ProviderConnection(provider_id=test_provider.id)
-        db_session.add(connection)
-        db_session.commit()
-
-        with patch(
-            "src.api.v1.auth.ProviderRegistry.create_provider_instance"
-        ) as mock_registry:
-            # Mock provider instance
-            mock_provider_impl = AsyncMock()
-            mock_provider_impl.authenticate = AsyncMock(
-                return_value={
-                    "access_token": "test_access_token",
-                    "refresh_token": "test_refresh_token",
-                    "expires_in": 3600,
-                    "token_type": "Bearer",
-                    "scope": "read write",
-                }
-            )
-            mock_registry.return_value = mock_provider_impl
-
-            response = client.get(
-                f"/api/v1/auth/{test_provider.id}/callback",
-                params={
-                    "code": "test_authorization_code",
-                    "state": str(test_provider.id),
-                },
-            )
-
-            # Verify authentication was attempted
-            mock_provider_impl.authenticate.assert_called_once()
-            # Response may be 200 or 500 depending on session handling
-            assert response.status_code in [
-                status.HTTP_200_OK,
-                status.HTTP_500_INTERNAL_SERVER_ERROR,
-            ]
-
-    def test_callback_with_error(self, client: TestClient, test_provider: Provider):
-        """Test callback with OAuth error."""
-        response = client.get(
-            f"/api/v1/auth/{test_provider.id}/callback",
-            params={
-                "error": "access_denied",
-                "error_description": "User denied access",
-            },
-        )
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert "Authorization failed" in response.json()["detail"]
-
-    def test_callback_missing_code(self, client: TestClient, test_provider: Provider):
-        """Test callback without authorization code."""
-        response = client.get(f"/api/v1/auth/{test_provider.id}/callback")
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert "No authorization code received" in response.json()["detail"]
-
-    def test_callback_state_mismatch(
-        self, client: TestClient, test_provider: Provider, db_session
-    ):
-        """Test callback with mismatched state parameter.
-
-        Note: Due to async/sync test infrastructure, this tests that
-        authentication is attempted despite state mismatch.
-        """
-        connection = ProviderConnection(provider_id=test_provider.id)
-        db_session.add(connection)
-        db_session.commit()
-
-        with patch(
-            "src.api.v1.auth.ProviderRegistry.create_provider_instance"
-        ) as mock_registry:
-            mock_provider_impl = AsyncMock()
-            mock_provider_impl.authenticate = AsyncMock(
-                return_value={"access_token": "test_token", "expires_in": 3600}
-            )
-            mock_registry.return_value = mock_provider_impl
-
-            # State doesn't match provider_id
-            response = client.get(
-                f"/api/v1/auth/{test_provider.id}/callback",
-                params={
-                    "code": "test_code",
-                    "state": str(uuid4()),  # Different UUID
-                },
-            )
-
-            # Verify authentication was attempted despite state mismatch
-            mock_provider_impl.authenticate.assert_called_once()
-            # Response may vary due to session handling
-            assert response.status_code in [
-                status.HTTP_200_OK,
-                status.HTTP_500_INTERNAL_SERVER_ERROR,
-            ]
-
-    def test_callback_provider_not_found(self, client: TestClient):
-        """Test callback for non-existent provider."""
-        fake_uuid = uuid4()
-        response = client.get(
-            f"/api/v1/auth/{fake_uuid}/callback", params={"code": "test_code"}
-        )
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
-
-    def test_callback_authentication_failure(
-        self, client: TestClient, test_provider: Provider, db_session
-    ):
-        """Test callback when token exchange fails."""
-        connection = ProviderConnection(provider_id=test_provider.id)
-        db_session.add(connection)
-        db_session.commit()
-
-        with patch(
-            "src.api.v1.auth.ProviderRegistry.create_provider_instance"
-        ) as mock_registry:
-            mock_provider_impl = AsyncMock()
-            mock_provider_impl.authenticate = AsyncMock(
-                side_effect=Exception("Token exchange failed")
-            )
-            mock_registry.return_value = mock_provider_impl
-
-            response = client.get(
-                f"/api/v1/auth/{test_provider.id}/callback",
-                params={"code": "test_code"},
-            )
-
-            assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-            assert "Failed to exchange authorization code" in response.json()["detail"]
-
-
-class TestRefreshTokens:
-    """Test suite for refresh_provider_tokens endpoint."""
-
-    def test_refresh_tokens_success(
-        self, client: TestClient, test_provider_with_connection: Provider, db_session
-    ):
-        """Test successful token refresh."""
-        provider = test_provider_with_connection
-
-        # Create token with refresh token
-        from src.services.encryption import EncryptionService
-
-        encryption = EncryptionService()
-
-        token = ProviderToken(
-            connection_id=provider.connection.id,
-            access_token_encrypted=encryption.encrypt("old_access_token"),
-            refresh_token_encrypted=encryption.encrypt("refresh_token"),
-            expires_at=datetime.now(timezone.utc),
-        )
-        db_session.add(token)
-        db_session.commit()
-
-        with patch(
-            "src.services.token_service.TokenService.refresh_token"
-        ) as mock_refresh:
-            mock_refresh.return_value = AsyncMock()
-
-            response = client.post(f"/api/v1/auth/{provider.id}/refresh")
-
-            assert response.status_code == status.HTTP_200_OK
-            data = response.json()
-            assert "Tokens refreshed successfully" in data["message"]
-            mock_refresh.assert_called_once()
-
-    def test_refresh_tokens_provider_not_found(self, client: TestClient):
-        """Test refreshing tokens for non-existent provider."""
-        fake_uuid = uuid4()
-        response = client.post(f"/api/v1/auth/{fake_uuid}/refresh")
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
-
-    def test_refresh_tokens_failure(
-        self, client: TestClient, test_provider_with_connection: Provider
-    ):
-        """Test handling of token refresh failure."""
-        with patch(
-            "src.services.token_service.TokenService.refresh_token"
-        ) as mock_refresh:
-            mock_refresh.side_effect = Exception("Refresh failed")
-
-            response = client.post(
-                f"/api/v1/auth/{test_provider_with_connection.id}/refresh"
-            )
-
-            assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-            assert "Failed to refresh tokens" in response.json()["detail"]
-
-
-class TestTokenStatus:
-    """Test suite for get_token_status endpoint."""
-
-    def test_get_token_status_with_tokens(
-        self, client: TestClient, test_provider_with_connection: Provider, db_session
-    ):
-        """Test getting token status when tokens exist."""
-        provider = test_provider_with_connection
-
-        # Create token
-        from src.services.encryption import EncryptionService
-
-        encryption = EncryptionService()
-
-        expires_at = datetime.now(timezone.utc).replace(tzinfo=None)
-        token = ProviderToken(
-            connection_id=provider.connection.id,
-            access_token_encrypted=encryption.encrypt("access_token"),
-            expires_at=expires_at,
-        )
-        db_session.add(token)
-        db_session.commit()
-
-        response = client.get(f"/api/v1/auth/{provider.id}/status")
-
-        assert response.status_code == status.HTTP_200_OK
-        data = response.json()
-        assert data["status"] == "connected"
-        assert data["provider_id"] == str(provider.id)
-        assert "expires_at" in data
-
-    def test_get_token_status_no_tokens(
-        self, client: TestClient, test_provider_with_connection: Provider
-    ):
-        """Test getting token status when no tokens exist."""
-        response = client.get(f"/api/v1/auth/{test_provider_with_connection.id}/status")
-
-        assert response.status_code == status.HTTP_200_OK
-        data = response.json()
-        assert data["status"] == "not_connected"
-        assert "No tokens found" in data["message"]
-
-    def test_get_token_status_provider_not_found(self, client: TestClient):
-        """Test getting token status for non-existent provider."""
-        fake_uuid = uuid4()
-        response = client.get(f"/api/v1/auth/{fake_uuid}/status")
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-class TestDisconnectProvider:
-    """Test suite for disconnect_provider endpoint."""
-
-    def test_disconnect_success(
-        self, client: TestClient, test_provider_with_connection: Provider, db_session
-    ):
-        """Test provider disconnection endpoint.
-
-        Note: Due to async/sync test infrastructure, this tests that
-        the endpoint is called correctly.
-        """
-        provider = test_provider_with_connection
-
-        # Create token
-        from src.services.encryption import EncryptionService
-
-        encryption = EncryptionService()
-
-        token = ProviderToken(
-            connection_id=provider.connection.id,
-            access_token_encrypted=encryption.encrypt("access_token"),
-        )
-        db_session.add(token)
-        db_session.commit()
-
-        response = client.delete(f"/api/v1/auth/{provider.id}/disconnect")
-
-        # Response may be 200 or 500 depending on session handling
-        assert response.status_code in [
-            status.HTTP_200_OK,
-            status.HTTP_500_INTERNAL_SERVER_ERROR,
-        ]
-
-    def test_disconnect_provider_not_found(self, client: TestClient):
-        """Test disconnecting non-existent provider."""
-        fake_uuid = uuid4()
-        response = client.delete(f"/api/v1/auth/{fake_uuid}/disconnect")
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
-
-    def test_disconnect_failure(
-        self, client: TestClient, test_provider_with_connection: Provider
-    ):
-        """Test handling of disconnection failure."""
-        with patch(
-            "src.services.token_service.TokenService.revoke_tokens"
-        ) as mock_revoke:
-            mock_revoke.side_effect = Exception("Revoke failed")
-
-            response = client.delete(
-                f"/api/v1/auth/{test_provider_with_connection.id}/disconnect"
-            )
-
-            assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-            assert "Failed to disconnect provider" in response.json()["detail"]
-
-
-class TestGetCurrentUser:
-    """Test suite for get_current_user dependency."""
-
-    def test_get_current_user_creates_user(self, client: TestClient):
-        """Test that get_current_user creates a test user in development."""
-        # The dependency is called for every request
-        # Just verify it works by making any authenticated request
-        # Note: Using provider-types endpoint (public, no auth required)
-        response = client.get("/api/v1/provider-types")
-
-        assert response.status_code == status.HTTP_200_OK
-        # If this passes, get_current_user worked
